+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `Attacher#assign_remote_url` which allows dynamically passing downloader options (@janko-m)
+
 * Deprecate `:storages` option in `download_endpoint` plugin in favour of `UploadedFile#download_url` (@janko-m)
 
 * Add `:redirect` option to `download_endpoint` plugin for redirecting to the uploaded file (@janko-m)